@@ -16,14 +16,14 @@ dependencies:
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
   provider: ^6.0.4
-  material_design_icons_flutter: ^6.0.7096
+  material_design_icons_flutter: ^7.0.7296
   firebase_core: ^2.3.0
   firebase_database: ^10.0.6
   firebase_auth: ^4.1.5
   shared_preferences: ^2.0.15
-  logger: ^1.1.0
+  logger: ^2.0.1
   bot_toast: ^4.0.2
-  image_picker: ^0.8.5+3
+  image_picker: ^1.0.1
   firebase_storage: ^11.0.6
   loading_animation_widget: ^1.2.0+2
   intl: ^0.18.0
@@ -33,11 +33,11 @@ dependencies:
   avatar_glow: ^2.0.2
   google_nav_bar: ^5.0.6
   iconsax: ^0.0.8
-  flutter_local_notifications: ^13.0.0
+  flutter_local_notifications: ^15.1.0+1
   sticky_grouped_list: ^3.1.0
   url_launcher: ^6.1.9
-  google_mobile_ads: ^2.3.0
-  package_info_plus: ^3.0.3
+  google_mobile_ads: ^3.0.0
+  package_info_plus: ^4.0.2
   json_annotation: ^4.8.0
 
 dev_dependencies:
