@@ -11,11 +11,13 @@ import 'package:portfolio_admin/widgets/custom_loading.dart';
 import 'package:portfolio_admin/widgets/custom_text_field.dart';
 import 'package:portfolio_admin/widgets/custom_toast.dart';
 import 'package:portfolio_admin/widgets/cutsom_app_bar.dart';
+import 'package:portfolio_admin/widgets/projects/custom_chip.dart';
 import 'package:provider/provider.dart';
 import 'package:url_launcher/url_launcher.dart';
 import '../../constants/colors.dart';
 import '../../models/project.dart';
-import '../../providers/works.dart';
+import '../../providers/skills.dart';
+import '../../widgets/drop_down.dart';
 import '../../widgets/projects/image_card.dart';
 
 class AddProjectScreen extends StatefulWidget {
@@ -40,7 +42,7 @@ class _AddProjectScreenState extends State<AddProjectScreen> {
   final ImagePicker _picker = ImagePicker();
   File? _coverImg;
   List<XFile> _pickedImages = [];
-
+  List<String> tech = [];
   Future<void> getCover() async {
     XFile? pickedImage = await _picker.pickImage(source: ImageSource.gallery);
     if (pickedImage == null) return;
@@ -48,6 +50,18 @@ class _AddProjectScreenState extends State<AddProjectScreen> {
     setState(() {
       _coverImg = imageTemp;
     });
+  }
+
+  @override
+  void initState() {
+    Future.delayed(Duration.zero).then((_) async {
+      try {
+        await Provider.of<Skills>(context, listen: false).fetchSkills();
+      } catch (e) {
+        Toast.showToast(message: e.toString(), type: ToastType.error);
+      }
+    });
+    super.initState();
   }
 
   @override
@@ -73,179 +87,202 @@ class _AddProjectScreenState extends State<AddProjectScreen> {
       floatingActionButton: _isLoading ? CustomLoading() : null,
       floatingActionButtonLocation: FloatingActionButtonLocation.centerFloat,
       appBar: CustomAppBar(title: 'Add Project', showLeading: true),
-      body: Container(
-        padding: const EdgeInsets.symmetric(horizontal: 15, vertical: 10),
-        child: Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            CustomTextField(
-              hint: 'Project Name',
-              preIcon: MdiIcons.pen,
-              tc: TextCapitalization.words,
-              data: _name,
-            ),
-            CustomTextField(
-              hint: 'Description',
-              preIcon: MdiIcons.paperclip,
-              tc: TextCapitalization.words,
-              data: _des,
-              lines: null,
-            ),
-            Text('Application Type', style: kBodyTextStyleGrey),
-            Row(
-              children: [
-                Radio(
-                  value: 0,
-                  groupValue: _groupValue,
-                  onChanged: (val) => setState(() => _groupValue = val!),
-                  activeColor: kPrimaryColor,
-                ),
-                Text('Mobile', style: kBodyTextStyleGrey),
-                const SizedBox(width: 10),
-                Radio(
-                  value: 1,
-                  groupValue: _groupValue,
-                  onChanged: (val) => setState(() => _groupValue = val!),
-                  activeColor: kPrimaryColor,
-                ),
-                Text('Web/Windows', style: kBodyTextStyleGrey),
-              ],
-            ),
-            Text('Images', style: kBodyTextStyleGrey),
-            Container(
-              margin: EdgeInsets.symmetric(vertical: 10),
-              child: Row(
-                crossAxisAlignment: CrossAxisAlignment.start,
+      body: SingleChildScrollView(
+        child: Container(
+          padding: const EdgeInsets.symmetric(horizontal: 15, vertical: 10),
+          child: Column(
+            crossAxisAlignment: CrossAxisAlignment.start,
+            children: [
+              CustomTextField(
+                hint: 'Project Name',
+                preIcon: MdiIcons.pen,
+                tc: TextCapitalization.words,
+                data: _name,
+              ),
+              CustomTextField(
+                hint: 'Description',
+                preIcon: MdiIcons.paperclip,
+                tc: TextCapitalization.words,
+                data: _des,
+                lines: null,
+              ),
+              Text('Application Type', style: kBodyTextStyleGrey),
+              Row(
                 children: [
-                  ImageCard(
-                    removeBgColor: _coverImg != null,
-                    click: getCover,
-                    child: _coverImg == null
-                        ? Column(
-                            mainAxisAlignment: MainAxisAlignment.center,
-                            children: [
-                              Icon(Icons.photo_camera_rounded,
-                                  color: kPrimaryColor),
-                              SizedBox(height: 5),
-                              Text(
-                                'Add Cover Image',
-                                style:
-                                    kBodyTextStyleGrey.copyWith(fontSize: 11),
-                                textAlign: TextAlign.center,
-                              ),
-                            ],
-                          )
-                        : Image.file(_coverImg!),
+                  Radio(
+                    value: 0,
+                    groupValue: _groupValue,
+                    onChanged: (val) => setState(() => _groupValue = val!),
+                    activeColor: kPrimaryColor,
                   ),
-                  SizedBox(width: 10),
-                  Visibility(
-                    visible: _pickedImages.isEmpty,
-                    child: ImageCard(
-                      click: () async {
-                        List<XFile>? pickedImages =
-                            await _picker.pickMultiImage();
-                        setState(() => _pickedImages = pickedImages);
-                      },
-                      child: Column(
-                        mainAxisAlignment: MainAxisAlignment.center,
-                        children: [
-                          Icon(Icons.photo_camera_rounded,
-                              color: kPrimaryColor),
-                          SizedBox(height: 5),
-                          Text(
-                            'Add More Images',
-                            style: kBodyTextStyleGrey.copyWith(fontSize: 11),
-                            textAlign: TextAlign.center,
-                          ),
-                        ],
-                      ),
-                    ),
+                  Text('Mobile', style: kBodyTextStyleGrey),
+                  const SizedBox(width: 10),
+                  Radio(
+                    value: 1,
+                    groupValue: _groupValue,
+                    onChanged: (val) => setState(() => _groupValue = val!),
+                    activeColor: kPrimaryColor,
                   ),
-                  Expanded(
-                    child: Wrap(
-                      runSpacing: 10.0,
-                      spacing: 10.0,
-                      children: _pickedImages
-                          .map((img) => SmallImageCard(image: img))
-                          .toList(),
-                    ),
-                  ),
+                  Text('Web/Windows', style: kBodyTextStyleGrey),
                 ],
               ),
-            ),
-            _groupValue == 1
-                ? CustomTextField(
-                    hint: 'External link',
-                    preIcon: MdiIcons.linkVariant,
-                    tc: TextCapitalization.none,
-                    data: _link,
-                    suffIcon: IconButton(
-                      onPressed: () async {
-                        Uri uri = Uri.parse(
-                            'https://app.netlify.com/teams/ericknamukolo/overview');
-                        await launchUrl(uri,
-                            mode: LaunchMode.externalApplication);
-                      },
-                      icon: Icon(Icons.link, color: kPrimaryColor),
+              Text('Images', style: kBodyTextStyleGrey),
+              Container(
+                margin: EdgeInsets.symmetric(vertical: 10),
+                child: Row(
+                  crossAxisAlignment: CrossAxisAlignment.start,
+                  children: [
+                    ImageCard(
+                      removeBgColor: _coverImg != null,
+                      click: getCover,
+                      child: _coverImg == null
+                          ? Column(
+                              mainAxisAlignment: MainAxisAlignment.center,
+                              children: [
+                                Icon(Icons.photo_camera_rounded,
+                                    color: kPrimaryColor),
+                                SizedBox(height: 5),
+                                Text(
+                                  'Add Cover Image',
+                                  style:
+                                      kBodyTextStyleGrey.copyWith(fontSize: 11),
+                                  textAlign: TextAlign.center,
+                                ),
+                              ],
+                            )
+                          : Image.file(_coverImg!),
                     ),
-                  )
-                : CustomTextField(
-                    hint: 'Google Playstore link',
-                    preIcon: MdiIcons.googlePlay,
-                    tc: TextCapitalization.none,
-                    data: _playStore,
-                    suffIcon: IconButton(
-                      onPressed: () async {
-                        Uri uri = Uri.parse(
-                            'https://play.google.com/store/apps/dev?id=8203990443766365712');
-                        await launchUrl(uri,
-                            mode: LaunchMode.externalApplication);
-                      },
-                      icon: Icon(Icons.link, color: kPrimaryColor),
+                    SizedBox(width: 10),
+                    Visibility(
+                      visible: _pickedImages.isEmpty,
+                      child: ImageCard(
+                        click: () async {
+                          List<XFile>? pickedImages =
+                              await _picker.pickMultiImage();
+                          setState(() => _pickedImages = pickedImages);
+                        },
+                        child: Column(
+                          mainAxisAlignment: MainAxisAlignment.center,
+                          children: [
+                            Icon(Icons.photo_camera_rounded,
+                                color: kPrimaryColor),
+                            SizedBox(height: 5),
+                            Text(
+                              'Add More Images',
+                              style: kBodyTextStyleGrey.copyWith(fontSize: 11),
+                              textAlign: TextAlign.center,
+                            ),
+                          ],
+                        ),
+                      ),
                     ),
-                  ),
-            CustomTextField(
-              hint: 'Github link',
-              preIcon: MdiIcons.github,
-              tc: TextCapitalization.none,
-              data: _github,
-              suffIcon: IconButton(
-                onPressed: () async {
-                  Uri uri = Uri.parse('https://github.com/ericknamukolo');
-                  await launchUrl(uri, mode: LaunchMode.externalApplication);
-                },
-                icon: Icon(Icons.link, color: kPrimaryColor),
+                    Expanded(
+                      child: Wrap(
+                        runSpacing: 10.0,
+                        spacing: 10.0,
+                        children: _pickedImages
+                            .map((img) => SmallImageCard(image: img))
+                            .toList(),
+                      ),
+                    ),
+                  ],
+                ),
               ),
-            ),
-            const Spacer(),
-            Consumer<Projects>(
-              builder: (context, data, _) => CustomButton(
-                btnText: 'Add Project',
-                isLoading: _isLoading,
-                click: () async {
-                  try {
-                    setState(() => _isLoading = true);
-                    await data.addProject(Project(
-                      name: _name.text,
-                      des: _des.text,
-                      type: appType[_groupValue],
-                      externalLink: _link.text,
-                      githubLink: _github.text,
-                      googleLink: _playStore.text,
-                      coverImg: _coverImg!,
-                      images: _pickedImages,
-                    ));
-                    setState(() => _isLoading = false);
-                    Navigator.of(context).pop();
-                  } catch (e) {
-                    Toast.showToast(
-                        message: e.toString(), type: ToastType.error);
-                    setState(() => _isLoading = false);
-                  }
-                },
+              Text('Tech Used', style: kBodyTextStyleGrey),
+              SizedBox(height: 5),
+              Consumer<Skills>(
+                builder: (_, value, __) => DropDown(
+                  items: value.skills.map((skill) => skill.name).toList(),
+                  hint: 'Select Tech Used',
+                  onChanged: (val) {
+                    if (!tech.contains(val)) {
+                      setState(() => tech.add(val!));
+                    }
+                  },
+                ),
               ),
-            ),
-          ],
+              Wrap(
+                runSpacing: 10.0,
+                children: tech
+                    .map((img) => CustomChip(
+                          name: img,
+                          delete: () => setState(() => tech.remove(img)),
+                        ))
+                    .toList(),
+              ),
+              Text('Links', style: kBodyTextStyleGrey),
+              SizedBox(height: 5),
+              CustomTextField(
+                hint: 'External link',
+                preIcon: MdiIcons.linkVariant,
+                tc: TextCapitalization.none,
+                data: _link,
+                suffIcon: IconButton(
+                  onPressed: () async {
+                    Uri uri = Uri.parse(
+                        'https://app.netlify.com/teams/ericknamukolo/overview');
+                    await launchUrl(uri, mode: LaunchMode.externalApplication);
+                  },
+                  icon: Icon(Icons.link, color: kPrimaryColor),
+                ),
+              ),
+              CustomTextField(
+                hint: 'Google Playstore link',
+                preIcon: MdiIcons.googlePlay,
+                tc: TextCapitalization.none,
+                data: _playStore,
+                suffIcon: IconButton(
+                  onPressed: () async {
+                    Uri uri = Uri.parse(
+                        'https://play.google.com/store/apps/dev?id=8203990443766365712');
+                    await launchUrl(uri, mode: LaunchMode.externalApplication);
+                  },
+                  icon: Icon(Icons.link, color: kPrimaryColor),
+                ),
+              ),
+              CustomTextField(
+                hint: 'Github link',
+                preIcon: MdiIcons.github,
+                tc: TextCapitalization.none,
+                data: _github,
+                suffIcon: IconButton(
+                  onPressed: () async {
+                    Uri uri = Uri.parse('https://github.com/ericknamukolo');
+                    await launchUrl(uri, mode: LaunchMode.externalApplication);
+                  },
+                  icon: Icon(Icons.link, color: kPrimaryColor),
+                ),
+              ),
+              Consumer<Projects>(
+                builder: (context, data, _) => CustomButton(
+                  btnText: 'Add Project',
+                  isLoading: _isLoading,
+                  click: () async {
+                    try {
+                      setState(() => _isLoading = true);
+                      await data.addProject(Project(
+                        name: _name.text,
+                        description: _des.text,
+                        type: appType[_groupValue],
+                        externalLink: _link.text,
+                        githubLink: _github.text,
+                        playstoreLink: _playStore.text,
+                        cover: _coverImg!,
+                        images: _pickedImages,
+                        tech: tech,
+                      ));
+                      setState(() => _isLoading = false);
+                      Navigator.of(context).pop();
+                    } catch (e) {
+                      Toast.showToast(
+                          message: e.toString(), type: ToastType.error);
+                      setState(() => _isLoading = false);
+                    }
+                  },
+                ),
+              ),
+            ],
+          ),
         ),
       ),
     );
